+++ R/mspeaklists-mzr.R
@@ -49,40 +49,38 @@ NULL
 #'
 #' @rdname MSPeakLists-generation
 #' @export
-generateMzRPeakLists <- function(fGroups, avgMzWindow = 0.001, maxRtMSWidth = NULL, avgTopPeaks = 50, avgMinIntensity = 500,
-                                 avgMassFun = mean, avgMethod = "distance", precursorMzWindow = 8, topMost = NULL)
+generateMzRPeakLists <- function(fGroups, maxRtMSWidth = 20, precursorMzWindow = 8, topMost = NULL,
+                                 avgFeatParams = getDefAvgPListParams(),
+                                 avgFGroupParams = getDefAvgPListParams())
 {
     ac <- checkmate::makeAssertCollection()
     checkmate::assertClass(fGroups, "featureGroups", add = ac)
-    aapply(checkmate::assertNumber, . ~ avgMzWindow + avgMinIntensity,
-           lower = 0, finite = TRUE, fixed = list(add = ac))
     checkmate::assertNumber(maxRtMSWidth, lower = 1, finite = TRUE, null.ok = TRUE, add = ac)
     checkmate::assertNumber(precursorMzWindow, lower = 0, finite = TRUE, null.ok = TRUE, add = ac)
-    checkmate::assertCount(avgTopPeaks, positive = TRUE, add = ac)
-    checkmate::assertFunction(avgMassFun, add = ac)
-    checkmate::assertChoice(avgMethod, c("distance", "hclust"), add = ac)
     checkmate::assertCount(topMost, positive = TRUE, null.ok = TRUE, add = ac)
+    assertAvgPListParams(avgFeatParams, add = ac)
+    assertAvgPListParams(avgFGroupParams, add = ac)
     checkmate::reportAssertions(ac)
     
     ftindex <- groupFeatIndex(fGroups)
     fTable <- featureTable(fGroups)
     anaInfo <- analysisInfo(fGroups)
     gTable <- groups(fGroups)
-    gcount <- ncol(ftindex)
+    gCount <- length(fGroups)
+    gNames <- names(fGroups)
     anaCount <- nrow(anaInfo)
     
-    if (gcount == 0)
+    if (gCount == 0)
         return(MSPeakLists(algorithm = "mzR"))
 
     cacheDB <- openCacheDBScope()
-    setHash <- makeHash(fGroups, avgMzWindow, maxRtMSWidth, avgTopPeaks, avgMinIntensity, avgMassFun, avgMethod,
-                        precursorMzWindow, topMost)
+    setHash <- makeHash(fGroups, maxRtMSWidth, precursorMzWindow, topMost, avgFeatParams)
     cachedSet <- loadCacheSet("MSPeakListsMzR", setHash, cacheDB)
-    resultHashes <- vector("character", anaCount * gcount)
+    resultHashes <- vector("character", anaCount * gCount)
     resultHashCount <- 0
 
     # structure: [[analysis]][[fGroup]][[MSType]][[MSPeak]]
-    ret <- list()
+    plists <- list()
 
     # if topMost is specified, make list (topAna) of topMost intense analyses
     if (!is.null(topMost) && topMost > 0)
@@ -94,18 +92,17 @@ generateMzRPeakLists <- function(fGroups, avgMzWindow = 0.001, maxRtMSWidth = NU
         fp <- getMzMLOrMzXMLAnalysisPath(ana, anaInfo$path[anai])
         spectra <- NULL
 
-        baseHash <- makeHash(ana, avgMzWindow, maxRtMSWidth, avgTopPeaks, avgMinIntensity, avgMassFun, avgMethod,
-                             precursorMzWindow, topMost)
+        baseHash <- makeHash(ana, maxRtMSWidth, precursorMzWindow, topMost, avgFeatParams)
 
-        printf("Loading all MS peak lists for %d feature groups in analysis '%s'...\n", gcount, ana)
-        prog <- txtProgressBar(0, gcount, style = 3)
+        printf("Loading all MS peak lists for %d feature groups in analysis '%s'...\n", gCount, ana)
+        prog <- txtProgressBar(0, gCount, style = 3)
 
         for (grpi in seq_along(ftindex))
         {
             if (!is.null(topMost) && !ana %in% topAna[[grpi]])
                 next # not intense enough
 
-            grp <- colnames(ftindex)[grpi]
+            grp <- gNames[grpi]
 
             fti <- ftindex[[grpi]][anai]
             if (fti == 0)
@@ -128,34 +125,34 @@ generateMzRPeakLists <- function(fGroups, avgMzWindow = 0.001, maxRtMSWidth = NU
 
                 rtRange <- c(ft$retmin, ft$retmax)
                 if (!is.null(maxRtMSWidth) && diff(rtRange) > maxRtMSWidth)
-                    rtRange <- c(max(rtRange[1], ft$rt - maxRtMSWidth/2), min(rtRange[2], ft$rt + maxRtMSWidth/2))
+                    rtRange <- c(max(rtRange[1], ft$ret - maxRtMSWidth/2), min(rtRange[2], ft$ret + maxRtMSWidth/2))
 
                 if (is.null(spectra))
                     spectra <- loadSpectra(fp, verbose = FALSE)
 
-                results$MS <- averageSpectra(spectra, rtRange, avgMzWindow, avgTopPeaks,
-                                             avgMinIntensity, avgMassFun = avgMassFun, method = avgMethod)
+                results$MS <- do.call(averageSpectraMZR, c(list(spectra, rtRange), avgFeatParams))
 
                 stopifnot(!is.null(ft$mz))
-                MSMS <- averageSpectra(spectra, rtRange, avgMzWindow, avgTopPeaks, avgMinIntensity, 2, ft$mz,
-                                       precursorMzWindow, avgMassFun, avgMethod)
+                MSMS <- do.call(averageSpectraMZR, c(list(spectra = spectra, rtRange = rtRange, MSLevel = 2,
+                                                          precursor = ft$mz, precursorMzWindow = precursorMzWindow),
+                                                     avgFeatParams))
                 if (nrow(MSMS) > 0)
                     results$MSMS <- MSMS
 
                 saveCacheData("MSPeakListsMzR", results, hash, cacheDB)
             }
 
-            ret[[ana]][[grp]] <- results
+            plists[[ana]][[grp]] <- results
 
             setTxtProgressBar(prog, grpi)
         }
 
-        setTxtProgressBar(prog, gcount)
+        setTxtProgressBar(prog, gCount)
         close(prog)
     }
 
     if (is.null(cachedSet))
-        saveCacheSet("MSPeakListsMzR", resultHashes[1:resultHashCount], setHash, cacheDB)
-
-    return(MSPeakLists(peakLists = ret, algorithm = "mzR"))
+        saveCacheSet("MSPeakListsMzR", resultHashes[seq_len(resultHashCount)], setHash, cacheDB)
+    
+    return(MSPeakLists(peakLists = plists, avgPeakListArgs = avgFGroupParams, algorithm = "mzR"))
 }